adding update and delete functions
