+++ src/main/java/com/autotune/common/exceptions/datasource/DataSourceNotExist.java
@@ -1,0 +1,10 @@
+package com.autotune.common.exceptions.datasource;
+
+public class DataSourceNotExist extends Exception {
+    public DataSourceNotExist() {
+    }
+
+    public DataSourceNotExist(String message) {
+        super(message);
+    }
+}
