+++ src/main/java/com/autotune/common/exceptions/datasource/DataSourceNotServiceable.java
@@ -1,4 +1,4 @@
-package com.autotune.common.exceptions;
+package com.autotune.common.exceptions.datasource;
 
 public class DataSourceNotServiceable extends Exception {
     public DataSourceNotServiceable() {
