+++ src/main/java/com/autotune/common/exceptions/datasource/KruizeDatasourceDoesNotExist.java
@@ -14,17 +14,18 @@
  * limitations under the License.
  *******************************************************************************/
 
-package com.autotune.common.exceptions;
+package com.autotune.common.exceptions.datasource;
 
 /**
- * KruizeDatasourceAlreadyExists is an exception which needs to be raised when there is
- * a duplicate entry trying to get added in the collection of datasources
+ * KruizeDatasourceDoesNotExist is a exception which needs to be raised when the datasource
+ * doesnot exist in the collection we search
  */
-public class KruizeDatasourceAlreadyExists extends Exception {
-    public KruizeDatasourceAlreadyExists() {
+
+public class KruizeDatasourceDoesNotExist extends Exception {
+    public KruizeDatasourceDoesNotExist() {
     }
 
-    public KruizeDatasourceAlreadyExists(String message) {
+    public KruizeDatasourceDoesNotExist(String message) {
         super(message);
     }
 }
