+++ src/main/java/com/autotune/common/exceptions/datasource/UnsupportedDataSourceProvider.java
@@ -1,4 +1,4 @@
-package com.autotune.common.exceptions;
+package com.autotune.common.exceptions.datasource;
 
 public class UnsupportedDataSourceProvider extends Exception {
     public UnsupportedDataSourceProvider() {
